+++ mullvad-daemon/src/lib.rs
@@ -305,8 +305,6 @@ pub enum DaemonCommand {
     #[cfg(daita)]
     SetEnableDaita(ResponseTx<(), settings::Error>, bool),
     #[cfg(daita)]
-    SetDaitaUseMultihopIfNecessary(ResponseTx<(), settings::Error>, bool),
-    #[cfg(daita)]
     SetDaitaSettings(ResponseTx<(), settings::Error>, DaitaSettings),
     /// Set DNS options or servers to use
     SetDnsOptions(ResponseTx<(), settings::Error>, DnsOptions),
@@ -1555,10 +1553,6 @@ impl Daemon {
             #[cfg(daita)]
             SetEnableDaita(tx, value) => self.on_set_daita_enabled(tx, value).await,
             #[cfg(daita)]
-            SetDaitaUseMultihopIfNecessary(tx, value) => {
-                self.on_set_daita_use_multihop_if_necessary(tx, value).await
-            }
-            #[cfg(daita)]
             SetDaitaSettings(tx, daita_settings) => {
                 self.on_set_daita_settings(tx, daita_settings).await
             }
@@ -2805,44 +2799,6 @@ impl Daemon {
             Err(e) => {
                 log::error!("{}", e.display_chain_with_msg("Unable to save settings"));
                 Self::oneshot_send(tx, Err(e), "set_daita_enabled response");
-            }
-        }
-    }
-
-    #[cfg(daita)]
-    async fn on_set_daita_use_multihop_if_necessary(
-        &mut self,
-        tx: ResponseTx<(), settings::Error>,
-        value: bool,
-    ) {
-        match self
-            .settings
-            .update(|settings| {
-                settings
-                    .tunnel_options
-                    .wireguard
-                    .daita
-                    .use_multihop_if_necessary = value
-            })
-            .await
-        {
-            Ok(settings_changed) => {
-                Self::oneshot_send(tx, Ok(()), "set_daita_use_multihop_if_necessary response");
-
-                if let RelaySettings::CustomTunnelEndpoint(_) = &self.settings.relay_settings {
-                    return; // DAITA is not supported for custom relays
-                }
-
-                let daita_enabled = self.settings.tunnel_options.wireguard.daita.enabled;
-
-                if settings_changed && daita_enabled {
-                    log::info!("Reconnecting because DAITA settings changed");
-                    self.reconnect_tunnel();
-                }
-            }
-            Err(e) => {
-                log::error!("{}", e.display_chain_with_msg("Unable to save settings"));
-                Self::oneshot_send(tx, Err(e), "set_daita_use_multihop_if_necessary response");
             }
         }
     }