+++ mullvad-management-interface/src/client.rs
@@ -308,12 +308,6 @@ impl MullvadProxyClient {
     }
 
     #[cfg(daita)]
-    pub async fn set_daita_direct_only(&mut self, value: bool) -> Result<()> {
-        self.0.set_daita_direct_only(value).await?;
-        Ok(())
-    }
-
-    #[cfg(daita)]
     pub async fn set_daita_settings(&mut self, settings: DaitaSettings) -> Result<()> {
         let settings = types::DaitaSettings::from(settings);
         self.0.set_daita_settings(settings).await?;